Rename pacdrive lights driver to win32
